added a listener for the initial data fetch
